+++ app/include/LLMSelectionDialog.hpp
@@ -35,6 +35,8 @@ public:
     std::string get_selected_custom_llm_id() const;
     std::string get_remote_api_key() const;
     std::string get_remote_model() const;
+    std::string get_gemini_api_key() const;
+    std::string get_gemini_model() const;
 
 private:
     void setup_ui();
@@ -43,7 +45,10 @@ private:
     void update_radio_selection();
     void update_custom_choice_ui();
     void update_remote_fields_state();
+    void update_gemini_fields_state();
     bool remote_inputs_valid() const;
+    bool gemini_inputs_valid() const;
+    bool validate_text_inputs(QLineEdit* key_edit, QLineEdit* model_edit) const;
     void update_local_choice_ui();
     void update_download_info();
     void start_download();
@@ -62,8 +67,11 @@ private:
     std::string selected_custom_id;
     std::string remote_api_key;
     std::string remote_model;
+    std::string gemini_api_key;
+    std::string gemini_model;
 
     QRadioButton* remote_radio{nullptr};
+    QRadioButton* gemini_radio{nullptr};
     QRadioButton* local3_radio{nullptr};
     QRadioButton* local7_radio{nullptr};
     QRadioButton* custom_radio{nullptr};
@@ -81,11 +89,17 @@ private:
     QDialogButtonBox* button_box{nullptr};
     QWidget* download_section{nullptr};
     QWidget* remote_inputs{nullptr};
+    QWidget* gemini_inputs{nullptr};
     QLineEdit* api_key_edit{nullptr};
     QLineEdit* model_edit{nullptr};
+    QLineEdit* gemini_api_key_edit{nullptr};
+    QLineEdit* gemini_model_edit{nullptr};
     QCheckBox* show_api_key_checkbox{nullptr};
+    QCheckBox* show_gemini_key_checkbox{nullptr};
     QLabel* remote_help_label{nullptr};
     QLabel* remote_link_label{nullptr};
+    QLabel* gemini_help_label{nullptr};
+    QLabel* gemini_link_label{nullptr};
 
     std::unique_ptr<LLMDownloader> downloader;
     std::atomic<bool> is_downloading{false};